+++ src/components/boto/boto.js
@@ -1,7 +1,17 @@
 import React from "react";
 
 
+const dirhola = () => console.log('hola');
 
 
 
+export const BotoSeguent = (props) => <button onClick={dirhola}>
+    {props.etiqueta}
+    
+  </button>
+
+
+export const BotoAnterior = (props) => <button>
+    {props.etiqueta}
+</button>
