+++ src/components/escena/escena.styled.js
@@ -5,7 +5,12 @@ import styled from "styled-components";
 
 
 
-
+export const EscenaStyled = styled.div`
+    border: 2px solid black;
+    border-radius: 40px;
+    margin: 25px;
+    
+`;
